+++ engine/include/cubos/engine/ui/canvas/canvas.hpp
@@ -1,12 +1,11 @@
 /// @file
 /// @brief Component @ref cubos::engine::UICanvas.
-/// @ingroup canvas-plugin
+/// @ingroup ui-canvas-plugin
 
 #pragma once
 
 #include <glm/glm.hpp>
 
-#include <cubos/core/gl/render_device.hpp>
 #include <cubos/core/reflection/reflect.hpp>
 
 #include <cubos/engine/api.hpp>
@@ -14,7 +13,7 @@
 namespace cubos::engine
 {
     /// @brief Component which represents a target for UI rendering.
-    /// @ingroup canvas-plugin
+    /// @ingroup ui-canvas-plugin
     struct CUBOS_ENGINE_API UICanvas
     {
         CUBOS_REFLECT;