+++ engine/include/cubos/engine/ui/canvas/horizontal_stretch.hpp
@@ -1,12 +1,9 @@
 /// @file
-/// @brief Component @ref cubos::engine::HorizontalStretch.
-/// @ingroup canvas-plugin
+/// @brief Component @ref cubos::engine::UIHorizontalStretch.
+/// @ingroup ui-canvas-plugin
 
 #pragma once
 
-#include <glm/glm.hpp>
-
-#include <cubos/core/gl/render_device.hpp>
 #include <cubos/core/reflection/reflect.hpp>
 
 #include <cubos/engine/api.hpp>
@@ -14,8 +11,8 @@
 namespace cubos::engine
 {
     /// @brief Component which makes a UI element fit horizontally to its parent.
-    /// @ingroup canvas-plugin
-    struct CUBOS_ENGINE_API HorizontalStretch
+    /// @ingroup ui-canvas-plugin
+    struct CUBOS_ENGINE_API UIHorizontalStretch
     {
         CUBOS_REFLECT;