+++ engine/include/cubos/engine/ui/canvas/plugin.hpp
@@ -1,9 +1,9 @@
 /// @dir
-/// @brief @ref canvas-plugin plugin directory.
+/// @brief @ref ui-canvas-plugin plugin directory.
 
 /// @file
 /// @brief Plugin entry point.
-/// @ingroup canvas-plugin
+/// @ingroup ui-canvas-plugin
 
 #pragma once
 
@@ -11,17 +11,28 @@
 
 namespace cubos::engine
 {
-    /// @defgroup canvas-plugin Canvas
+    /// @defgroup ui-canvas-plugin Canvas
     /// @ingroup engine
     /// @brief Adds UI canvas and elements.
 
+    /// @brief @ref UICanvas components' information is passed to related @ref UIElement components.
+    /// @ingroup canvas-target-plugin
     CUBOS_ENGINE_API extern Tag elementVPUpdateTag;
+
+    /// @brief The @ref UIElement components are updated with information from their parents.
+    /// @ingroup canvas-target-plugin
     CUBOS_ENGINE_API extern Tag elementPropagateTag;
+
+    /// @brief Readies canvas for drawing.
+    /// @ingroup canvas-target-plugin
     CUBOS_ENGINE_API extern Tag uiBeginTag;
+
+    /// @brief Systems which draw to canvas should be tagged with this.
+    /// @ingroup canvas-target-plugin
     CUBOS_ENGINE_API extern Tag uiDrawTag;
 
     /// @brief Plugin entry function.
     /// @param cubos @b CUBOS. main class
-    /// @ingroup canvas-plugin
-    CUBOS_ENGINE_API void canvasPlugin(Cubos& cubos);
+    /// @ingroup ui-canvas-plugin
+    CUBOS_ENGINE_API void uiCanvasPlugin(Cubos& cubos);
 } // namespace cubos::engine