+++ engine/src/ui/canvas/horizontal_stretch.cpp
@@ -1,16 +1,13 @@
 
 #include <cubos/core/ecs/reflection.hpp>
-#include <cubos/core/reflection/external/glm.hpp>
 #include <cubos/core/reflection/external/primitives.hpp>
 
 #include <cubos/engine/ui/canvas/horizontal_stretch.hpp>
 
-using namespace cubos::core::gl;
-
-CUBOS_REFLECT_IMPL(cubos::engine::HorizontalStretch)
+CUBOS_REFLECT_IMPL(cubos::engine::UIHorizontalStretch)
 {
-    return cubos::core::ecs::TypeBuilder<HorizontalStretch>("cubos::engine::HorizontalStretch")
-        .withField("leftOffset", &HorizontalStretch::leftOffset)
-        .withField("rightOffset", &HorizontalStretch::rightOffset)
+    return cubos::core::ecs::TypeBuilder<UIHorizontalStretch>("cubos::engine::UIHorizontalStretch")
+        .withField("leftOffset", &UIHorizontalStretch::leftOffset)
+        .withField("rightOffset", &UIHorizontalStretch::rightOffset)
         .build();
 }