+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- UI canvas, elements, and color rects. (#1116, **@DiogoMendonc-a**)
+
 ## [v0.2.0] - 2024-05-07
 
 ### Added